requirements.txt
beautifulsoup4==4.12.3
fake_useragent==1.4.0
httpx==0.27.2
Pillow==11.0.0
pycryptodome==3.20.0
pytesseract==0.3.10
Requests==2.32.3
toml==0.10.2
